Conflicts:
	src/emu/bus/cpc/hd20.c
	src/emu/cpu/hphybrid/hphybrid.c
	src/emu/cpu/hphybrid/hphybrid.h
	src/emu/machine/netlist.h
	src/emu/netlist/devices/net_lib.c
	src/emu/netlist/devices/nld_truthtable.c
	src/emu/netlist/nl_base.c
	src/emu/netlist/nl_base.h
	src/emu/netlist/nl_setup.h
	src/emu/netlist/plib/plists.h
	src/emu/netlist/solver/nld_ms_direct.h
	src/emu/netlist/solver/nld_ms_gmres.h
	src/emu/netlist/tools/nl_convert.c
	src/emu/video/rgbsse.h
	src/emu/video/rgbvmx.h
	src/emu/video/vooddefs.h
	src/mame/audio/irem.c
	src/mame/audio/nl_kidniki.c
	src/mame/drivers/aleisttl.c
	src/mame/drivers/cclimber.c
	src/mame/drivers/megasys1.c
	src/mame/drivers/namcos86.c
	src/mame/includes/cclimber.h
	src/mame/includes/ladybug.h
	src/mame/machine/m1comm.c
	src/mame/video/segas32.c
	src/mess/drivers/hp64k.c
	src/mess/drivers/i7000.c
	src/mess/drivers/rc759.c
	src/version.c

+++ src/emu/mewui/selector.c
@@ -42,7 +42,7 @@ void ui_menu_selector::handle()
 	{
 		if (menu_event->iptkey == IPT_UI_SELECT)
 		{
-			for (int idx = 0; idx < str_items.size(); idx++)
+			for (size_t idx = 0; idx < str_items.size(); idx++)
 				if ((void*)&str_items[idx] == menu_event->itemref)
 					*selector = idx;
 
@@ -121,7 +121,7 @@ void ui_menu_selector::populate()
 	}
 	else
 	{
-		for (int index = 0, added = 0; index < str_items.size(); index++)
+		for (size_t index = 0, added = 0; index < str_items.size(); index++)
 			if (str_items[index].compare("_skip_") != 0)
 			{
 				if (first_pass && *selector == index)
@@ -149,7 +149,7 @@ void ui_menu_selector::custom_render(void *selectedref, float top, float bottom,
 
 	// get the size of the text
 	machine().ui().draw_text_full(container, tempbuf.c_str(), 0.0f, 0.0f, 1.0f, JUSTIFY_CENTER, WRAP_TRUNCATE,
-									DRAW_NONE, ARGB_WHITE, ARGB_BLACK, &width, NULL);
+	                              DRAW_NONE, ARGB_WHITE, ARGB_BLACK, &width, NULL);
 	width += (2.0f * UI_BOX_LR_BORDER) + 0.01f;
 	float maxwidth = MAX(width, origx2 - origx1);
 
@@ -169,7 +169,7 @@ void ui_menu_selector::custom_render(void *selectedref, float top, float bottom,
 
 	// draw the text within it
 	machine().ui().draw_text_full(container, tempbuf.c_str(), x1, y1, x2 - x1, JUSTIFY_CENTER, WRAP_TRUNCATE,
-									DRAW_NORMAL, UI_TEXT_COLOR, UI_TEXT_BG_COLOR, NULL, NULL);
+	                              DRAW_NORMAL, UI_TEXT_COLOR, UI_TEXT_BG_COLOR, NULL, NULL);
 
 	// bottom text
 	// get the text for 'UI Select'
@@ -178,7 +178,7 @@ void ui_menu_selector::custom_render(void *selectedref, float top, float bottom,
 	tempbuf.assign("Double click or press ").append(ui_select_text.c_str()).append(" to select");
 
 	machine().ui().draw_text_full(container, tempbuf.c_str(), 0.0f, 0.0f, 1.0f, JUSTIFY_CENTER, WRAP_NEVER,
-									DRAW_NONE, ARGB_WHITE, ARGB_BLACK, &width, NULL);
+	                              DRAW_NONE, ARGB_WHITE, ARGB_BLACK, &width, NULL);
 	width += 2 * UI_BOX_LR_BORDER;
 	maxwidth = MAX(maxwidth, width);
 
@@ -198,7 +198,7 @@ void ui_menu_selector::custom_render(void *selectedref, float top, float bottom,
 
 	// draw the text within it
 	machine().ui().draw_text_full(container, tempbuf.c_str(), x1, y1, x2 - x1, JUSTIFY_CENTER, WRAP_NEVER,
-									DRAW_NORMAL, UI_TEXT_COLOR, UI_TEXT_BG_COLOR, NULL, NULL);
+	                              DRAW_NORMAL, UI_TEXT_COLOR, UI_TEXT_BG_COLOR, NULL, NULL);
 }
 
 //-------------------------------------------------
